tests/wpt/tests/html/semantics/forms/form-submission-0/form-submission-blocked-by-sandboxing.html

<!DOCTYPE html>
<html>
  <link rel="help" href="https://html.spec.whatwg.org/multipage/#concept-form-submit">
  <title>Test that form submission is blocked by a document's active sandboxing flags</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <body>
  <html>
    <form id="f">
      <input type="hidden" value="test" />
    </form>
    <script>
      async_test((t) => {
        var form = document.getElementById('f')

        form.addEventListener('submit', t.step_func_done((e) => {
          e.preventDefault()
          assert_true(false, 'form should not be submitted due to sandboxing flags');
        }));

        form.submit();
        t.step_timeout(() => t.done(), 500);
      }, 'csp-derived sandboxing flags prevent form submission.')
    </script>
</body>
</html>
